feat: add comment module

+++ src/modules/users/users.service.ts
@@ -3,6 +3,7 @@ import {
   HttpCode,
   Injectable,
   NotFoundException,
+  ForbiddenException,
 } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
@@ -49,24 +50,32 @@ export class UsersService {
   }
 
   async findAds(id: string) {
-    console.log(id)
     return await this.usersRepository.findAds(id);
   }
 
-  async update(id: string, updateUserDto: UpdateUserDto) {
+  async update(id: string, updateUserDto: UpdateUserDto, req_id: string) {
     const user = await this.usersRepository.findOne(id);
     if (!user) {
       throw new NotFoundException('User not found.');
     }
-    const updateUser = await this.usersRepository.update(id, updateUserDto);
-    return updateUser;
+    if (id != req_id) {
+      throw new ForbiddenException(
+        'Not allowed to update other users. You can patch just your own profile.',
+      );
+    }
+    return await this.usersRepository.update(id, updateUserDto);
   }
 
   @HttpCode(204)
-  async remove(id: string) {
+  async remove(id: string, req_id: string) {
     const user = await this.usersRepository.findOne(id);
     if (!user) {
       throw new NotFoundException('User not found.');
+    }
+    if (id != req_id) {
+      throw new ForbiddenException(
+        'Not allowed to remove other users. You can delete just your own profile.',
+      );
     }
     return await this.usersRepository.delete(id);
   }
@@ -86,16 +95,16 @@ export class UsersService {
       user.name,
       resetToken,
     );
-    await this.mailService.sendEmail(resetPasswordTemplate)
+    await this.mailService.sendEmail(resetPasswordTemplate);
   }
 
   @HttpCode(200)
-  async resetPassword(password: string, reset_token: string){
+  async resetPassword(password: string, reset_token: string) {
     const user = await this.usersRepository.findByToken(reset_token);
     if (!user) {
       throw new NotFoundException('User not found.');
     }
 
-    await this.usersRepository.updatePassword(user.id, password)
+    await this.usersRepository.updatePassword(user.id, password);
   }
 }
